Account registration page fully functional.
Some additions:
Upon registration, the user will be redirected back to the login page with a flash message that their account has been registered and they can log in. Also if the email submitted in the registration page has already been used in the past, a flash warning will be shown, and the user must submit a different email.