added pc case gear scraper and removed new egg scraping
added pc case gear scraping to the newScraper.py file and removed the new egg scraping from it. Also added a chrome driver application to the code base, this driver may not run on your device depending on what chrome version you have. The imports required for the drive are included at the top and will require a selenium install.